bugprone: make char/integer conversions explicit and safe
  - signed-char-misuse: FASTQ quality symbols (a signed 'char') were converted
    to int/size_t for score-table indexing directly. A malformed quality byte
    >127 would become negative and, as a size_t subscript, index far out of
    bounds. Cast through 'unsigned char' first (mergepairs quality merging and
    ee accumulation; derep quality merge) -- behaviour-identical for valid
    input (33-126) and safe for the rest.
  - narrowing-conversions: make the implementation-defined narrowings explicit:
    static_cast<short> on the SIMD gap-penalty sums (16-bit-guarded), the
    unsigned->int kmer-count shift in kmerhash, and the unsigned->char lowercase
    result in mask.

+++ src/core/align_simd.cpp
@@ -1579,23 +1579,23 @@ auto search16(s16info_s * s,
 
   R_query_left = v_dup(s->penalty_gap_extension_query_left);
 
-  QR_query_interior = v_dup((s->penalty_gap_open_query_interior +
+  QR_query_interior = v_dup(static_cast<short>(s->penalty_gap_open_query_interior +
                              s->penalty_gap_extension_query_interior));
   R_query_interior  = v_dup(s->penalty_gap_extension_query_interior);
 
-  QR_query_right  = v_dup((s->penalty_gap_open_query_right +
+  QR_query_right  = v_dup(static_cast<short>(s->penalty_gap_open_query_right +
                            s->penalty_gap_extension_query_right));
   R_query_right  = v_dup(s->penalty_gap_extension_query_right);
 
-  QR_target_left  = v_dup((s->penalty_gap_open_target_left +
+  QR_target_left  = v_dup(static_cast<short>(s->penalty_gap_open_target_left +
                            s->penalty_gap_extension_target_left));
   R_target_left  = v_dup(s->penalty_gap_extension_target_left);
 
-  QR_target_interior = v_dup((s->penalty_gap_open_target_interior +
+  QR_target_interior = v_dup(static_cast<short>(s->penalty_gap_open_target_interior +
                               s->penalty_gap_extension_target_interior));
   R_target_interior = v_dup(s->penalty_gap_extension_target_interior);
 
-  QR_target_right  = v_dup((s->penalty_gap_open_target_right +
+  QR_target_right  = v_dup(static_cast<short>(s->penalty_gap_open_target_right +
                             s->penalty_gap_extension_target_right));
   R_target_right  = v_dup(s->penalty_gap_extension_target_right);
 #pragma GCC diagnostic pop
